+++ libs/remix-ui/tree-view/src/types/index.ts
@@ -1,6 +1,4 @@
 /* eslint-disable no-undef */
-import { FileType } from 'libs/remix-ui/workspace/src/lib/types/index'
-
 export interface TreeViewProps {
     children?: React.ReactNode,
     id?: string
@@ -27,5 +25,5 @@ export interface TreeViewItemProps {
     showIcon?: boolean
     expandedPaths?: string[];
     handleClickFolder?: (path: string, type: string) => void;
-    file?: FileType
+    file?: any
 }
